refactor: Move search bar into its own component
